golf_tracker/requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.3
filterpy==1.4.5
pillow==10.1.0
flask==3.0.0
